feat: user can now remove the location filter (#127)
* feat: user can now remove the location filter

When authenticated the user received a location which filtered the prices and couldnt be removed.
Now the location is a filter in the form of a chip and can be removed. Doing so the user has access
to all prices.

* test: memoize lib some tsconfigs

to use the memoize function the allowSyntheticDefaultImports flag has to be enabled. For the tests,
enabling the esModuleInterop is necessary